+++ apps/dbagent/src/components/chat/artifacts/artifact-close-button.tsx
