+++ apps/dbagent/src/components/chat/artifacts/artifact.tsx
@@ -8,15 +8,15 @@ import { AnimatePresence, motion } from 'framer-motion';
 import { memo, useCallback, useEffect, useState } from 'react';
 import { useDebounceCallback, useWindowSize } from 'usehooks-ts';
 import { Document, Vote } from '~/lib/db/schema';
+import { MultimodalInput } from '../multimodal-input';
+import { fetcher } from '../utils';
 import { ArtifactActions } from './artifact-actions';
 import { ArtifactCloseButton } from './artifact-close-button';
 import { ArtifactMessages } from './artifact-messages';
-import { sheetArtifact } from './artifacts/sheet/client';
-import { textArtifact } from './artifacts/text/client';
-import { MultimodalInput } from './multimodal-input';
+import { sheetArtifact } from './sheet/client';
+import { textArtifact } from './text/client';
 import { Toolbar } from './toolbar';
 import { useArtifact } from './use-artifact';
-import { fetcher } from './utils';
 import { VersionFooter } from './version-footer';
 
 export const artifactDefinitions = [textArtifact, sheetArtifact];
