+++ apps/dbagent/src/components/chat/artifacts/data-stream-handler.tsx
