+++ apps/dbagent/src/components/chat/artifacts/document-skeleton.tsx
