+++ apps/dbagent/src/components/chat/artifacts/sheet/client.tsx
@@ -2,8 +2,8 @@ import { toast } from '@internal/components';
 import { CopyIcon, SparklesIcon, UndoIcon } from 'lucide-react';
 import { parse, unparse } from 'papaparse';
 import { RedoIcon } from '~/components/icons';
-import { Artifact } from '../../create-artifact';
-import { SpreadsheetEditor } from '../../sheet-editor';
+import { Artifact } from '../create-artifact';
+import { SpreadsheetEditor } from './editor';
 
 type Metadata = any;
