+++ apps/dbagent/src/components/chat/artifacts/text/actions.ts
