+++ apps/dbagent/src/components/chat/artifacts/text/editor/config.ts
