+++ apps/dbagent/src/components/chat/message/index.tsx
@@ -8,9 +8,9 @@ import { AnimatePresence, motion } from 'framer-motion';
 import { Clock, SparklesIcon } from 'lucide-react';
 import { memo, useState } from 'react';
 import { Vote } from '~/lib/db/schema';
-import { PencilEditIcon } from '../icons';
-import { DocumentToolCall, DocumentToolResult } from './document';
-import { DocumentPreview } from './document-preview';
+import { PencilEditIcon } from '../../icons';
+import { DocumentToolCall, DocumentToolResult } from '../artifacts/document';
+import { DocumentPreview } from '../artifacts/document-preview';
 import { Markdown } from './markdown';
 import { MessageActions } from './message-actions';
 import { MessageEditor } from './message-editor';
